requirements.txt
aes==1.0.0
asgiref==3.3.4
blinker==1.4
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
click==7.1.2
cryptography==3.2.1
Flask==1.1.4
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
kaitaistruct==0.9
ldap3==2.8.1
MarkupSafe==2.0.1
mitmproxy==5.3.0
msgpack==1.0.3
passlib==1.7.4
protobuf==3.13.0
publicsuffix2==2.20191221
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
six==1.16.0
sortedcontainers==2.2.2
tornado==6.1
typing_extensions==4.0.0
urwid==2.1.2
Werkzeug==1.0.1
wsproto==0.15.0
zstandard==0.14.1
